+++ src/components/dashboard/CourseCarousel.tsx
@@ -1,10 +1,12 @@
 
-import React from 'react';
+import React, { useState, useCallback, useEffect } from 'react';
+import { ChevronLeft, ChevronRight } from 'lucide-react';
+import { EmblaOptionsType } from 'embla-carousel';
+import useEmblaCarousel from 'embla-carousel-react';
+import { Button } from '@/components/ui/button';
 import CourseCardWithProgress from './CourseCardWithProgress';
 import { Podcast } from '@/types';
-import { useCourseCarousel } from './carousel/useCourseCarousel';
-import CourseCarouselNavigationArrows from './carousel/CourseCarouselNavigationArrows';
-import CourseCarouselScrollIndicators from './carousel/CourseCarouselScrollIndicators';
+import { useIsMobile } from '@/hooks/use-mobile';
 
 interface CourseCarouselProps {
   title: string;
@@ -28,16 +30,57 @@ const CourseCarousel: React.FC<CourseCarouselProps> = ({
   onToggleSave,
   onCourseClick
 }) => {
-  const {
-    emblaRef,
-    emblaApi,
-    canScrollPrev,
-    canScrollNext,
-    selectedIndex,
-    scrollPrev,
-    scrollNext,
-    isMobile
-  } = useCourseCarousel({ coursesLength: courses.length });
+  const isMobile = useIsMobile();
+  const [canScrollPrev, setCanScrollPrev] = useState(false);
+  const [canScrollNext, setCanScrollNext] = useState(false);
+
+  // Unified Embla configuration matching TouchCarousel
+  const options: EmblaOptionsType = {
+    align: 'start',
+    slidesToScroll: 1,
+    containScroll: 'trimSnaps',
+    dragFree: false, // Changed for better touch control
+    skipSnaps: false, // Ensure proper stopping at cards
+  };
+
+  const [emblaRef, emblaApi] = useEmblaCarousel(options);
+
+  const scrollPrev = useCallback(() => {
+    if (emblaApi) {
+      console.log('CourseCarousel: Scrolling to previous slide');
+      emblaApi.scrollPrev();
+    }
+  }, [emblaApi]);
+
+  const scrollNext = useCallback(() => {
+    if (emblaApi) {
+      console.log('CourseCarousel: Scrolling to next slide');
+      emblaApi.scrollNext();
+    }
+  }, [emblaApi]);
+
+  const onSelect = useCallback(() => {
+    if (!emblaApi) return;
+    setCanScrollPrev(emblaApi.canScrollPrev());
+    setCanScrollNext(emblaApi.canScrollNext());
+    console.log('CourseCarousel: Carousel state updated', {
+      canScrollPrev: emblaApi.canScrollPrev(),
+      canScrollNext: emblaApi.canScrollNext()
+    });
+  }, [emblaApi]);
+
+  useEffect(() => {
+    if (!emblaApi) return;
+    console.log('CourseCarousel: Embla API initialized');
+    onSelect();
+    emblaApi.on('select', onSelect);
+    emblaApi.on('reInit', onSelect);
+
+    return () => {
+      emblaApi.off('select', onSelect);
+      emblaApi.off('reInit', onSelect);
+    };
+  }, [emblaApi, onSelect]);
 
   if (courses.length === 0) {
     return (
@@ -50,32 +93,30 @@ const CourseCarousel: React.FC<CourseCarouselProps> = ({
     );
   }
 
+  // Unified navigation logic
+  const shouldShowNavigation = isMobile ? courses.length > 1 : courses.length > 4;
+
   return (
     <div className="mb-6 sm:mb-8">
       <h2 className="text-xl sm:text-2xl font-bold mb-3 sm:mb-4 px-4 sm:px-0">{title}</h2>
       
       <div className="relative px-4 sm:px-0">
+        {/* Unified carousel structure with touch optimization */}
         <div className="relative">
           <div 
-            className="overflow-hidden cursor-grab active:cursor-grabbing" 
+            className="overflow-hidden touch-pan-x" 
             ref={emblaRef}
-            style={{ 
-              touchAction: 'pan-x pinch-zoom',
-              WebkitOverflowScrolling: 'touch'
-            }}
+            style={{ touchAction: 'pan-x' }}
           >
-            <div className="flex touch-pan-x">
+            <div className="flex">
               {courses.map((course) => (
                 <div 
                   key={course.podcast.id} 
                   className={`flex-none ${
                     isMobile 
-                      ? 'w-[280px] min-w-[280px] mr-4 first:ml-0 last:mr-0' 
+                      ? 'w-[85vw] max-w-[320px] mr-4 first:ml-0 last:mr-0' 
                       : 'w-full sm:w-1/2 lg:w-1/3 xl:w-1/4 pr-6'
                   }`}
-                  style={{ 
-                    flex: isMobile ? '0 0 280px' : undefined
-                  }}
                 >
                   <CourseCardWithProgress
                     podcast={course.podcast}
@@ -92,22 +133,33 @@ const CourseCarousel: React.FC<CourseCarouselProps> = ({
             </div>
           </div>
 
-          <CourseCarouselNavigationArrows
-            isMobile={isMobile}
-            coursesLength={courses.length}
-            canScrollPrev={canScrollPrev}
-            canScrollNext={canScrollNext}
-            scrollPrev={scrollPrev}
-            scrollNext={scrollNext}
-          />
-        </div>
+          {/* Navigation arrows - unified logic */}
+          {shouldShowNavigation && (
+            <>
+              <Button
+                variant="outline"
+                size="icon"
+                className="absolute left-0 top-1/2 -translate-y-1/2 -translate-x-2 sm:-translate-x-4 bg-white shadow-lg z-10 h-8 w-8 sm:h-10 sm:w-10"
+                onClick={scrollPrev}
+                disabled={!canScrollPrev}
+              >
+                <ChevronLeft className="h-3 w-3 sm:h-4 sm:w-4" />
+              </Button>
 
-        <CourseCarouselScrollIndicators
-          isMobile={isMobile}
-          coursesLength={courses.length}
-          selectedIndex={selectedIndex}
-          emblaApi={emblaApi}
-        />
+              <Button
+                variant="outline"
+                size="icon"
+                className="absolute right-0 top-1/2 -translate-y-1/2 translate-x-2 sm:translate-x-4 bg-white shadow-lg z-10 h-8 w-8 sm:h-10 sm:w-10"
+                onClick={scrollNext}
+                disabled={!canScrollNext}
+              >
+                <ChevronRight className="h-3 w-3 sm:h-4 sm:w-4" />
+              </Button>
+            </>
+          )}
+        </div>
+        
+        {/* Pagination dots removed - no longer needed */}
       </div>
     </div>
   );
