+++ src/components/dashboard/TouchCarousel.tsx
@@ -1,10 +1,12 @@
 
-import React from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
+import { EmblaOptionsType } from 'embla-carousel';
+import useEmblaCarousel from 'embla-carousel-react';
+import { ChevronLeft, ChevronRight } from 'lucide-react';
+import { Button } from '@/components/ui/button';
 import CourseCardWithProgress from './CourseCardWithProgress';
 import { Podcast } from '@/types';
-import { useTouchCarousel } from './carousel/useTouchCarousel';
-import CarouselNavigationArrows from './carousel/CarouselNavigationArrows';
-import CarouselScrollIndicators from './carousel/CarouselScrollIndicators';
+import { useIsMobile } from '@/hooks/use-mobile';
 
 interface TouchCarouselProps {
   title: string;
@@ -28,16 +30,57 @@ const TouchCarousel: React.FC<TouchCarouselProps> = ({
   onToggleSave,
   onCourseClick
 }) => {
-  const {
-    emblaRef,
-    emblaApi,
-    canScrollPrev,
-    canScrollNext,
-    selectedIndex,
-    scrollPrev,
-    scrollNext,
-    isMobile
-  } = useTouchCarousel({ coursesLength: courses.length });
+  const isMobile = useIsMobile();
+  const [canScrollPrev, setCanScrollPrev] = useState(false);
+  const [canScrollNext, setCanScrollNext] = useState(false);
+
+  // Optimized Embla configuration for better touch handling
+  const options: EmblaOptionsType = {
+    align: 'start',
+    slidesToScroll: 1,
+    containScroll: 'trimSnaps',
+    dragFree: false, // Changed from true to false for better control
+    skipSnaps: false, // Ensure it stops at cards properly
+  };
+
+  const [emblaRef, emblaApi] = useEmblaCarousel(options);
+
+  const scrollPrev = useCallback(() => {
+    if (emblaApi) {
+      console.log('TouchCarousel: Scrolling to previous slide');
+      emblaApi.scrollPrev();
+    }
+  }, [emblaApi]);
+
+  const scrollNext = useCallback(() => {
+    if (emblaApi) {
+      console.log('TouchCarousel: Scrolling to next slide');
+      emblaApi.scrollNext();
+    }
+  }, [emblaApi]);
+
+  const onSelect = useCallback(() => {
+    if (!emblaApi) return;
+    setCanScrollPrev(emblaApi.canScrollPrev());
+    setCanScrollNext(emblaApi.canScrollNext());
+    console.log('TouchCarousel: Carousel state updated', {
+      canScrollPrev: emblaApi.canScrollPrev(),
+      canScrollNext: emblaApi.canScrollNext()
+    });
+  }, [emblaApi]);
+
+  useEffect(() => {
+    if (!emblaApi) return;
+    console.log('TouchCarousel: Embla API initialized');
+    onSelect();
+    emblaApi.on('select', onSelect);
+    emblaApi.on('reInit', onSelect);
+
+    return () => {
+      emblaApi.off('select', onSelect);
+      emblaApi.off('reInit', onSelect);
+    };
+  }, [emblaApi, onSelect]);
 
   if (courses.length === 0) {
     return (
@@ -50,33 +93,29 @@ const TouchCarousel: React.FC<TouchCarouselProps> = ({
     );
   }
 
+  // Determine when to show navigation arrows
+  const shouldShowNavigation = isMobile ? courses.length > 1 : courses.length > 4;
+
   return (
     <div className="mb-8 sm:mb-12">
       <h2 className="text-xl sm:text-2xl font-bold mb-6 px-4 sm:px-0">{title}</h2>
       
       <div className="relative">
-        {/* Mobile-optimized carousel container with snap */}
+        {/* Unified carousel structure with optimized touch CSS */}
         <div 
-          className="overflow-hidden cursor-grab active:cursor-grabbing" 
+          className="overflow-hidden touch-pan-x" 
           ref={emblaRef}
-          style={{ 
-            touchAction: 'pan-x pinch-zoom',
-            WebkitOverflowScrolling: 'touch'
-          }}
+          style={{ touchAction: 'pan-x' }}
         >
-          <div className={`flex ${isMobile ? 'snap-x snap-mandatory' : ''}`}>
+          <div className="flex">
             {courses.map((course) => (
               <div 
                 key={course.podcast.id} 
                 className={`flex-none ${
                   isMobile 
-                    ? 'w-[90vw] max-w-[320px] px-4 snap-center snap-always' 
+                    ? 'w-[85vw] max-w-[320px] mr-4 first:ml-4 last:mr-4' 
                     : 'w-full sm:w-1/2 lg:w-1/3 xl:w-1/4 pr-6'
                 }`}
-                style={{ 
-                  flex: isMobile ? '0 0 90vw' : undefined,
-                  maxWidth: isMobile ? '320px' : undefined
-                }}
               >
                 <CourseCardWithProgress
                   podcast={course.podcast}
@@ -93,21 +132,30 @@ const TouchCarousel: React.FC<TouchCarouselProps> = ({
           </div>
         </div>
 
-        <CarouselNavigationArrows
-          isMobile={isMobile}
-          coursesLength={courses.length}
-          canScrollPrev={canScrollPrev}
-          canScrollNext={canScrollNext}
-          scrollPrev={scrollPrev}
-          scrollNext={scrollNext}
-        />
+        {/* Navigation arrows - unified logic */}
+        {shouldShowNavigation && (
+          <>
+            <Button
+              variant="outline"
+              size="icon"
+              className="absolute left-0 top-1/2 -translate-y-1/2 -translate-x-2 sm:-translate-x-4 bg-white shadow-lg z-10 h-8 w-8 sm:h-10 sm:w-10"
+              onClick={scrollPrev}
+              disabled={!canScrollPrev}
+            >
+              <ChevronLeft className="h-3 w-3 sm:h-4 sm:w-4" />
+            </Button>
 
-        <CarouselScrollIndicators
-          isMobile={isMobile}
-          coursesLength={courses.length}
-          selectedIndex={selectedIndex}
-          emblaApi={emblaApi}
-        />
+            <Button
+              variant="outline"
+              size="icon"
+              className="absolute right-0 top-1/2 -translate-y-1/2 translate-x-2 sm:translate-x-4 bg-white shadow-lg z-10 h-8 w-8 sm:h-10 sm:w-10"
+              onClick={scrollNext}
+              disabled={!canScrollNext}
+            >
+              <ChevronRight className="h-3 w-3 sm:h-4 sm:w-4" />
+            </Button>
+          </>
+        )}
       </div>
     </div>
   );
